Convert from worker eval to temp file import
Avoid dealing with all the nonsense of making the worker work.

Specifically: The worker evaluates in the context of the current working directory, not the package, so it may fail to resolve dependencies of build-mdx-to-html (e.g. React), depending how NPM chooses to dedupe the package tree. I could not sort out how to force it to resolve as though it were in the package.